tp/pyproject.toml
[tool.poetry]
name = "tp3"
version = "0.1.0"
description = "ITBA-TP3"
authors = ["Florencia Riccomi"]

[tool.poetry.dependencies]
python = "^3.8"
pip = "^21.3.1"
psycopg2-binary = "^2.9.1"
pandas = "1.1.0"
python-decouple = "^3.5"
requests = "^2.26.0"
apache-airflow = "^2.2.4"
apache-airflow-providers-postgres = "^4.0.0"
matplotlib = "^3.5.1"
SQLAlchemy="1.3.18"
aiobotocore = "^2.1.2"
boto3 = "1.20.0"
fsspec = "^2022.2.0"
s3fs = "^2022.2.0"
apache-airflow-providers-amazon = "^3.2.0"
sagemaker = "1.72.1"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
flake8 = "^4.0.1"
copilot = "^0.1.9"
black = {version = "^21.10b0", allow-prereleases = true}
pdbpp = "^0.10.3"
mypy = "^0.910"
memory-profiler = "^0.58.0"
Sphinx = "^4.3.0"
flake8-docstrings = "^1.6.0"
pytest = "^6.2.5" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
